+++ service/config_provider.go
@@ -81,26 +81,7 @@ type configProvider struct {
 // * Then unmarshalls the final config.Config using the given configunmarshaler.ConfigUnmarshaler.
 //
 // The `configMapProviders` is a map of pairs <scheme,Provider>.
-//
-// Notice: This API is experimental.
 func MustNewConfigProvider(
-	locations []string,
-	configMapProviders map[string]configmapprovider.Provider,
-	cfgMapConverters []config.MapConverterFunc,
-	configUnmarshaler configunmarshaler.ConfigUnmarshaler) ConfigProvider {
-	return NewConfigProvider(locations, configMapProviders, cfgMapConverters, configUnmarshaler)
-}
-
-// Deprecated: [v0.44.0] use MustNewConfigProvider instead
-// NewConfigProvider returns a new ConfigProvider that provides the configuration:
-// * Retrieve the config.Map by merging all retrieved maps from all the configmapprovider.Provider in order.
-// * Then applies all the config.MapConverterFunc in the given order.
-// * Then unmarshalls the final config.Config using the given configunmarshaler.ConfigUnmarshaler.
-//
-// The `configMapProviders` is a map of pairs <scheme,Provider>.
-//
-// Notice: This API is experimental.
-func NewConfigProvider(
 	locations []string,
 	configMapProviders map[string]configmapprovider.Provider,
 	cfgMapConverters []config.MapConverterFunc,
@@ -120,14 +101,7 @@ func NewConfigProvider(
 // MustNewDefaultConfigProvider returns the default ConfigProvider, and it creates configuration from a file
 // defined by the given configFile and overwrites fields using properties.
 func MustNewDefaultConfigProvider(configLocations []string, properties []string) ConfigProvider {
-	return NewDefaultConfigProvider(configLocations, properties)
-}
-
-// Deprecated: [v.0.44.0] use MustNewDefaultConfigProvider instead
-// NewDefaultConfigProvider returns the default ConfigProvider, and it creates configuration from a file
-// defined by the given configFile and overwrites fields using properties.
-func NewDefaultConfigProvider(configLocations []string, properties []string) ConfigProvider {
-	return NewConfigProvider(
+	return MustNewConfigProvider(
 		configLocations,
 		map[string]configmapprovider.Provider{
 			"file": configmapprovider.NewFile(),